Move creatable template types to /models
