+++ src/functions/templates/create-one-template.ts
@@ -6,14 +6,14 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { NumericID, RawNumericID, toRawNumericID, UUID } from '../../value-objects';
+import { CreatableTemplate, toRawCreatableTemplate } from '../../models';
+import { UUID } from '../../value-objects';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
 	buildURL,
 	fetch,
 	HTTPRequestOptions,
-	omitUndefinedShallow,
 	parseJSONResponse,
 } from '../utils';
 
@@ -38,64 +38,3 @@ export const makeCreateOneTemplate = (makerOptions: APIFunctionMakerOptions) => 
 		}
 	};
 };
-
-export interface CreatableTemplate {
-	userID?: NumericID;
-	groupIDs?: Array<NumericID>;
-
-	name: string;
-	description?: string | null;
-	labels?: Array<string>;
-
-	isGlobal?: boolean;
-	isRequired: boolean;
-
-	query: string;
-	variable: {
-		token: string;
-		name: string;
-		description?: string | null;
-	};
-
-	previewValue?: string | null;
-}
-
-interface RawCreatableTemplate {
-	Name: string;
-	Description: string | null;
-	Contents: {
-		query: string;
-		variable: string;
-		variableLabel: string;
-		variableDescription: string | null;
-		required: boolean;
-		testValue: string | null;
-	};
-
-	// !WARNING: That's not working right now, CHECK
-	UID?: RawNumericID;
-	GIDs: Array<RawNumericID>;
-
-	Global: boolean;
-	Labels: Array<string>;
-}
-
-const toRawCreatableTemplate = (creatable: CreatableTemplate): RawCreatableTemplate =>
-	omitUndefinedShallow({
-		UID: creatable.userID ? toRawNumericID(creatable.userID) : undefined,
-		GIDs: creatable.groupIDs?.map(id => toRawNumericID(id)) ?? [],
-
-		Global: creatable.isGlobal ?? false,
-		Labels: creatable.labels ?? [],
-
-		Name: creatable.name,
-		Description: creatable.description ?? null,
-		Contents: {
-			required: creatable.isRequired,
-			query: creatable.query,
-			variable: creatable.variable.token,
-			variableDescription: creatable.variable.description ?? null,
-			variableLabel: creatable.variable.name,
-			testValue: creatable.previewValue ?? null,
-		},
-	});
