+++ src/functions/templates/get-all-templates.spec.ts
@@ -6,13 +6,13 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { UUID } from '../../value-objects';
-import { isTemplate } from '../../models';
+import { CreatableTemplate, isTemplate } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
-import { CreatableTemplate, makeCreateOneTemplate } from './create-one-template';
-import { makeGetAllTemplates } from './get-all-templates';
+import { UUID } from '../../value-objects';
+import { makeCreateOneTemplate } from './create-one-template';
 import { makeDeleteOneTemplate } from './delete-one-template';
+import { makeGetAllTemplates } from './get-all-templates';
 
 describe('getAllTemplates()', () => {
 	const getAllTemplates = makeGetAllTemplates({ host: TEST_HOST, useEncryption: false });
