+++ src/models/template/index.ts
@@ -6,7 +6,10 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
+export * from './creatable-template';
 export * from './is-template';
+export * from './raw-creatable-template';
 export * from './raw-template';
 export * from './template';
+export * from './to-raw-creatable-template';
 export * from './to-template';
